Added Worm.script
Worm.script is a worm-like script created for BitBurner that spreads from server to server indefinitely unless all scripts are killed. It will also do the basic functions of a weaken/grow/hack script. If you have your own script for this, you can simply remove that part and replace it with code to copy and run your hacking script. However, you must make sure that you make room in the RAM for both the hacking script and the worm so that they both have room to run.